requirements.txt
numpy
tqdm
tensorflow
pillow
ffmpeg_python
torch
torchvision
scikit-learn
pyYAML
matplotlib
timm
